Increase deadline of momentum_sgd_test to 1000 ms
When HIP first launches the kernel it takes more time since it needs to read the kernel from shared library. If
there are lots of kernels it may take hundreds of ms. This test uses hypothesis which has a default deadline
of 200ms. If any test run exceeds the deadline the run will be flagged as fail. It seems the default deadline
is not sufficient.