Use -rpath-link to fix libtinfo conflict.
The ROCm libtinfo6 is conflicting with the libtinfo6
installed in the conda envirnonment in /opt/conda/lib.
This is first in the build rpath and causes issues on
Ubuntu 20.04 when resolving libamd_comgr.so
dependencies. Use -rpath-link to resolve dependencies
where the system libtinfo is found instead. This is
currently active on versions >= 20.04.