[release/2.4] Skipped test_slice_mm_bandwidth_computation and test_pad_addmm_dyn_mn on navi (#1674)
The tests fail because is_big_gpu for navi returns false and GEMMs
backend fails to produce any kernels going back to ATen. The tests fail
to check to valid metadate for kernels and thus fail. Decided to skip
the for navi currently and not rely on is_big_gpu check.