[Release/2..4] Fix 493250 tests - make barrier() block cpu and pass two unit tests (#1714)
Fixes
#[SWDEV-493250](https://ontrack-internal.amd.com/browse/SWDEV-493250)

Added @with_nccl_blocking_wait decorator to tests
test_barrier_full_group_cuda and test_barrier_group_cuda so that the CPU
blocks until the GPU streams are synchronized in the barrier call and
the assertion check is made after the GPU streams are synchronized.
barrier() call is made to block CPU in the upstream after release2.4,
but this functionality is not present in the upstream release2.4. This
decorator helps to make barrier() block the CPU and pass these unit
tests for rocm/pytorch release2.4.